requirements.txt
ccxt>=4.5.0
fastapi>=0.135.0
uvicorn>=0.44.0
psycopg2-binary>=2.9.0
